refactor: move `renderHighlighedTitle` into lib/util
